Extract and fix remaining handles tests
