+++ src/lib.rs
@@ -10,13 +10,17 @@ mod unittest_batch_burn;
 mod unittest_batch_mint;
 mod unittest_batch_send;
 mod unittest_batch_transfer;
-mod unittest_handles;
+mod unittest_contract_initialization;
 mod unittest_burn;
-mod unittest_cw721_approve;
-mod unittest_cw721_revoke;
-mod unittest_helpers;
-mod unittest_mint;
+mod unittest_check_permissions;
 mod unittest_contract_admin;
+mod unittest_cw721_approve;
+mod unittest_cw721_approve_all;
+mod unittest_cw721_revoke;
+mod unittest_cw721_revoke_all;
+mod unittest_helpers;
+mod unittest_make_ownership_private;
+mod unittest_mint;
 mod unittest_private_metadata;
 mod unittest_public_metadata;
 mod unittest_queries;
@@ -24,6 +28,7 @@ mod unittest_reveal;
 mod unittest_transfer;
 mod unittest_register_receive_nft;
 mod unittest_send;
+mod unittest_set_global_approval;
 mod unittest_viewing_key;
 mod unittest_whitelisted_approval;
 mod utils;
